+++ src/services/document.service.ts
@@ -1,16 +1,24 @@
 import axios from "axios";
 import { getAccessToken } from "./auth.service";
+import { DocumentResponseDto } from "../types/types";
 
 const API_URL = process.env.NEXT_PUBLIC_API_URL; // Use environment variables for base URL
 
 export const createDocument = async (
   caseId: string,
   mediaId: string,
-  title: string
+  title: string,
+  type: string,
+  mainDocumentId: string
 ) => {
   const response = await axios.post(
     `${API_URL}/cases/${caseId}/documents`,
-    { title, mediaId },
+    {
+      title: title,
+      mediaId: mediaId,
+      mainDocumentId: mainDocumentId,
+      type: type,
+    },
     {
       headers: {
         Authorization: `Bearer ${getAccessToken()}`,
@@ -20,13 +28,3 @@ export const createDocument = async (
 
   return response.data as DocumentResponseDto;
 };
-export interface DocumentResponseDto {
-  id: string;
-  title: string;
-  mediaId: string;
-  mediaUrl: string;
-  createdAt: string;
-  processingStatus: string;
-  citationsExtractionStatus: string | null;
-  citationsCount: number;
-}
