+++ src/utils/util.constants.ts
@@ -1,0 +1,4 @@
+export const DocType = {
+    MAIN: 'Main',
+    EXHIBIT: 'Exhibit',
+}
